backend/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python-headless
pillow
numpy
python-multipart